fix: remove test line
fix: remove qe.init

feat: remove qr msg if cancelled

feat: keepdays and dayspac supports float

+++ src/qzone3tg/app/storage/__init__.py
@@ -168,7 +168,7 @@ class DefaultStorageHook(StorageEvent):
             if flush:
                 await sess.commit()
 
-    def add_clean_task(self, keepdays: int, interval: float = 86400):
+    def add_clean_task(self, keepdays: float, interval: float = 86400):
         """
         This function register a timer that calls `self.clean(-keepdays * 86400)`
         every `interval` seconds.
